\name{aws.s3-package}
\alias{aws.s3-package}
\alias{aws.s3}
\docType{package}
\title{AWS S3 Client Package}
\description{A simple client package for the Amazon Web Services (AWS) Simple Storage Service (S3) REST API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
