\name{aws.signature-package}
\alias{aws.signature-package}
\alias{aws.signature}
\docType{package}
\title{Amazon Web Services Request Signatures}
\description{Generates Amazon Web Services (AWS) request signatures for RESTful APIs.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
