% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{bucketexists}
\alias{bucketexists}
\title{Check whether a bucket exists and is accessible with the current authentication keys.}
\usage{
bucketexists(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if bucket exists and is accessible, else \code{FALSE}.
An \code{aws_error} object may be returned if the request failed.
}
\description{
Check whether a bucket exists and is accessible with the current authentication keys.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketHEAD.html}{API Documentation}
}

