% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{delete_replication}
\alias{delete_replication}
\title{Delete the replication policy for a bucket.}
\usage{
delete_replication(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise.
An \code{aws_error} object may be returned if the request failed.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketDELETEreplication.html}{API Documentation}
}

