% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getobject}
\alias{getobject}
\alias{saveobject}
\alias{headobject}
\alias{copyobject}
\alias{copybucket}
\alias{putbucket}
\alias{putobject}
\alias{deleteobject}
\alias{getbucket}
\alias{deletebucket}
\alias{bucketexists}
\title{Deprecated}
\usage{
getobject(...)

saveobject(...)

headobject(...)

copyobject(...)

copybucket(...)

putbucket(...)

putobject(...)

deleteobject(...)

getbucket(...)

deletebucket(...)

bucketexists(...)
}
\arguments{
\item{\dots}{Arguments passed to updated versions of each function.}
}
\description{
These functions are deprecated.
}
