% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_base_url}
\alias{get_base_url}
\title{Get the default base url}
\usage{
get_base_url()
}
\value{
the default base url.
}
\description{
get the default location for all aw3.s3
request. By default, the base url is "s3.amazonaws.com"
}
