% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_bucket}
\alias{get_bucket}
\title{Get bucket}
\usage{
get_bucket(bucket, prefix = NULL, delimiter = NULL, max = NULL,
  marker = NULL, parse_response = TRUE, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{prefix}{Character string that limits the response to keys that begin 
with the specified prefix}

\item{delimiter}{Character string used to group keys.  Read the AWS doc for more detail.}

\item{max}{Integer indicating the maximum number of keys to return (max 1000).}

\item{marker}{Character string that pecifies the key to start with when 
listing objects in a bucket. Amazon S3 returns object keys in alphabetical order, 
starting with key after the marker in order.}

\item{parse_response}{logical, should we attempt to parse the response?}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list of objects in the bucket. If \code{parse_response = FALSE}, a nested list with the
complete contents of the AWS response.
}
\description{
List the contents of an S3 bucket
}
\details{
from the AWS doc: \dQuote{This implementation of the GET operation returns some
 or all (up to 1000) of the objects in a bucket. You can use the request parameters 
 as selection criteria to return a subset of the objects in a bucket.}
}
\references{
\href{https://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGET.html}{API Documentation}
}

