% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_cors}
\alias{get_cors}
\title{Get the cross origin resource sharing configuration information for a bucket.}
\usage{
get_cors(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list with cors configuration and rules.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGETcors.html}{API Documentation}
}

