% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_lifecycle}
\alias{get_lifecycle}
\title{Get the lifecycle configuration information for a bucket.}
\usage{
get_lifecycle(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list with lifecycle configuration, if it has been configured.
}
\description{
Get the lifecycle configuration information for a bucket.
}

