% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_replication}
\alias{get_replication}
\title{Get the replication configuration for a bucket.}
\usage{
get_replication(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the replication configuration, if one has been set.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGETreplication.html}{API Documentation}
}

