% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_requestpayment}
\alias{get_requestpayment}
\title{Get the requestPayment subresource for a bucket.}
\usage{
get_requestpayment(bucket, ...)
}
\arguments{
\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the requestPayment information, if set.
}
\description{
Get the requestPayment subresource for a bucket.
}
\examples{
\dontrun{}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTrequestPaymentGET.html}{API Documentation}
}
\keyword{bucket}

