% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_tagging}
\alias{get_tagging}
\title{Get the tag set for a bucket.}
\usage{
get_tagging(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the tag set, if one has been set.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGETtagging.html}{API Documentation}
}

