% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{get_torrent}
\alias{get_torrent}
\title{Retrieves a Bencoded dictionary (BitTorrent) for an object from an S3 bucket.}
\usage{
get_torrent(bucket, object, ...)
}
\arguments{
\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{object}{A character string containing the name of an object, or an object of class \dQuote{s3_object}.}

\item{...}{additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
Something.
}
\description{
Retrieves a Bencoded dictionary (BitTorrent) for an object from an S3 bucket.
}
\examples{
\dontrun{}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGETtorrent.html}{API Documentation}
}
\keyword{object}

