% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_versioning}
\alias{get_versioning}
\title{Get the versioning status of a bucket.}
\usage{
get_versioning(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket you want to get.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}}
}
\value{
If versioning has never been enabled or suspend, the value is \code{NULL}.
Otherwise, the status is returned (either \dQuote{Enabled} or \dQuote{Suspended}).
}
\description{
Get the versioning status of a bucket.
}

