% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_versions}
\alias{get_versions}
\title{Get versions of bucket objects.}
\usage{
get_versions(bucket, ...)
}
\arguments{
\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list.
}
\description{
Get versions of bucket objects.
}
\examples{
\dontrun{}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGETVersion.html}{API Documentation}
}
\keyword{bucket}

