% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{getobject}
\alias{getobject}
\title{Retrieves an object from an S3 bucket}
\usage{
getobject(bucket, object, headers = list(), ...)
}
\arguments{
\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{object}{A character string containing the name of an object, or an object of class \dQuote{s3_object}.}

\item{headers}{List of request headers for the REST call.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A raw object.
}
\description{
Retrieves an object from an S3 bucket
}
\examples{
\dontrun{}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation}
}
\keyword{object}

