% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{get_notification}
\alias{get_notification}
\alias{put_notification}
\title{Notifications}
\usage{
get_notification(bucket, ...)

put_notification(bucket, request_body, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}

\item{request_body}{A character string containing an XML request body, as defined in the specification in the \href{https://docs.aws.amazon.com/AmazonS3/latest/API/API_PutBucketNotificationConfiguration.html}{API Documentation}.}
}
\value{
A list containing the notification configuration, if one has been set.
}
\description{
Get/put the notification configuration for a bucket.
}
\references{
\href{https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetBucketNotificationConfiguration.html}{API Documentation: GET}
\href{https://docs.aws.amazon.com/AmazonS3/latest/API/API_PutBucketNotificationConfiguration.html}{API Documentation: PUT}
}
