% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{get_policy}
\alias{delete_policy}
\alias{get_policy}
\alias{put_policy}
\title{Bucket policies}
\usage{
get_policy(bucket, ...)

put_policy(bucket, ...)

delete_policy(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
For \code{get_policy}: A list with a policy, if one has been set.. For \code{delete_policy}: \code{TRUE} if successful, \code{FALSE} otherwise. An \code{aws_error} object may be returned if the request failed.
}
\description{
Get/Put/Delete the bucket access policy for a bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGETpolicy.html}{API Documentation}
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketDELETEpolicy.html}{API Documentation}
}

