% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{put_versioning}
\alias{put_versioning}
\title{Set the versioning status of a bucket.}
\usage{
put_versioning(bucket, status = c("Enabled", "Suspended"), ...)
}
\arguments{
\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{status}{Character string specifying whether versioning should be
\dQuote{Enabled} or \dQuote{Suspended}.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
If versioning has never been enabled or suspend, the value is \code{NULL}.
Otherwise, the status is returned (either \dQuote{Enabled} or \dQuote{Suspended}).
}
\description{
Set the versioning status of a bucket.
}
\examples{
\dontrun{}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketPUTVersioningStatus.html}{API Documentation}
}
\keyword{bucket}

