% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{putbucket}
\alias{putbucket}
\title{Creates a new S3 bucket.}
\usage{
putbucket(bucket, ...)
}
\arguments{
\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, aws_error details if not.
}
\description{
Creates a new S3 bucket.
}
\examples{
\dontrun{}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketPUT.html}{API Documentation}
}
\keyword{bucket}

