% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{putobject}
\alias{putobject}
\title{puts an object into a s3 bucket}
\usage{
putobject(bucket, object, ...)
}
\arguments{
\item{bucket}{Character string of the name of the bucket you want to get.}

\item{object}{Character string of the name of the object you want to get.}

\item{...}{additional arguments passed to \code{\link{s3HTTP}}}
}
\value{
list, containing aws api response
}
\description{
puts an object into a s3 bucket
}

