% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestPayment.R
\name{get_requestpayment}
\alias{get_requestpayment}
\alias{put_requestpayment}
\title{requestPayment}
\usage{
get_requestpayment(bucket, ...)

put_requestpayment(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the requestPayment information, if set.
}
\description{
Get/Put the requestPayment subresource for a bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTrequestPaymentGET.html}{API Documentation}
}
