% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_path.R
\name{read_s3_path}
\alias{read_s3_path}
\title{Treatment of the s3 path.}
\usage{
read_s3_path(s3_path)
}
\arguments{
\item{path}{For \code{s3_path}, a single path that would be decompose into object and bucket.}
}
\value{
For \code{s3_path}, a vector with two variables, first one is bucket second one is object.
}
\description{
Use S3 path to return object and bucket without needing further coding.
}
\examples{
\dontrun{
s3read_using(read_delim, delim = ";", object = s3_path(path)[2],bucket = s3_path(path[1]))
}
}
