% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3curl.R
\name{s3curl}
\alias{s3curl}
\title{s3curl}
\usage{
s3curl(object, bucket, open = "rb", accelerate = FALSE,
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"))
}
\arguments{
\item{object}{Character string with the object key, or an object of class \dQuote{s3_object}. In most cases, if \code{object} is specified as the latter, \code{bucket} can be omitted because the bucket name will be extracted from \dQuote{Bucket} slot in \code{object}.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{open}{One of \dQuote{r} or \dQuote{rb}.}

\item{accelerate}{A logical indicating whether to use AWS transfer acceleration, which can produce significant speed improvements for cross-country transfers. Acceleration only works with buckets that do not have dots in bucket name.}

\item{region}{A character string containing the AWS region. Ignored if region can be inferred from \code{bucket}.
If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. 
If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key. 
If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}
}
\value{
If \code{file = NULL}, a raw object. Otherwise, a character string containing the file name that the object is saved to.
}
\description{
Create an s3curl connection
}
\details{
This function creates a \code{\link[curl]{curl}} connection to be used for asynchronous reading from an S3 object.
}
\seealso{
\code{\link{get_object}}, \code{\link{save_object}}, \code{\link[curl]{curl}}
}
