% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http.r
\name{s3HTTP}
\alias{s3HTTP}
\title{S3 HTTP Requests}
\usage{
s3HTTP(verb = "GET", bucket = "", path = "", headers = list(),
  request_body = "", region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), parse_response = TRUE, ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{bucket}{Character string with the name of the bucket.}

\item{path}{A character string with the name of the object to put in the bucket
(sometimes called the object or 'key name' in the AWS documentation.)}

\item{headers}{a list of request headers for the REST call.}

\item{request_body}{character string of request body data.}

\item{region}{A character string containing the AWS region.
If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID.
If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key.
If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}

\item{parse_response}{return the response as is, or parse and return as a list?
default is TRUE.}

\item{...}{Additional arguments passed to an HTTP request function,
such as \code{\link[httr]{GET}}.}
}
\value{
the S3 response, or the relevant error.
}
\description{
This is the workhorse function for executing API requests for S3.
}
\details{
This is mostly an internal function for executing API requests.
In almost all cases, users do not need to access this directly.
}
\author{
Thomas J. Leeper \email{thosjleeper@gmail.com}
}

