% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveload.r
\name{s3load}
\alias{s3load}
\title{Load from S3}
\usage{
s3load(bucket, object, opts, envir = parent.frame())
}
\arguments{
\item{bucket}{A character string containing the name of the bucket you want
to delete an object from.}

\item{object}{A character string containing the name of the object.}

\item{opts}{An optional list of options to be passed to \code{\link{getobject}}.}

\item{envir}{An R environment, by default the parent environment, into which the R
object(s) should be loaded}
}
\value{
If successful, \code{NULL}, else the contents of the API error.
}
\description{
Load an R object from S3 ala \code{\link[base]{load}}.
}
\details{
Load an R object directly from an S3 bucket rather than a local directory.
}
\examples{
\dontrun{}
}
\seealso{
\code{\link{getobject}}, \code{\link{s3save}}
}
\keyword{bucket,}
\keyword{object}

