% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveload.r
\name{s3save}
\alias{s3save}
\title{Save to S3}
\usage{
s3save(..., bucket, object, opts = list())
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}

\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{object}{A character string containing the name of an object, or an object of class \dQuote{s3_object}.}

\item{opts}{An optional list of options to be passed to \code{\link{postobject}}.}
}
\value{
If successful, \code{NULL}, else the contents of the API error.
}
\description{
Save an R object to S3 ala \code{\link[base]{save}}.
}
\details{
Save an R object directly to an S3 bucket rather than a local directory.
}
\examples{
\dontrun{}
}
\seealso{
\code{\link{postobject}}, \code{\link{s3load}}
}
\keyword{bucket,}
\keyword{object}

