% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3saveRDS.R
\name{s3saveRDS}
\alias{s3saveRDS}
\alias{s3readRDS}
\title{saveRDS/readRDS}
\usage{
s3saveRDS(x, bucket, object = paste0(as.character(substitute(x)), ".rds"),
  compress = TRUE, ...)

s3readRDS(bucket, object, ...)
}
\arguments{
\item{x}{For \code{s3saveRDS}, a single R object to be saved via \code{\link[base]{saveRDS}} and uploaded to S3. \code{x} is analogous to the \code{object} argument in \code{saveRDS}.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{object}{Character string with the object key, or an object of class \dQuote{s3_object}. In most cases, if \code{object} is specified as the latter, \code{bucket} can be omitted because the bucket name will be extracted from \dQuote{Bucket} slot in \code{object}.}

\item{compress}{A logical. See \code{\link[base]{saveRDS}}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
For \code{s3saveRDS}, a logical. For \code{s3readRDS}, an R object.
}
\description{
Serialization interface to read/write R objects to S3
}
\examples{
\dontrun{
# create bucket
b <- put_bucket("myexamplebucket")

# save a single object to s3
s3saveRDS(x = mtcars, bucket = "myexamplebucket", object = "mtcars.rds")

# restore it under a different name
mtcars2 <- s3readRDS(object = "mtcars.rds", bucket = "myexamplebucket")
identical(mtcars, mtcars2)

# cleanup
delete_object(object = "mtcars.rds", bucket = "myexamplebucket")
delete_bucket("myexamplebucket")
}
}
\seealso{
\code{\link{s3save}},\code{\link{s3load}}
}
\author{
Steven Akins <skawesome@gmail.com>
}
