% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_base_url}
\alias{set_base_url}
\title{Set the default base url}
\usage{
set_base_url(new_url)
}
\arguments{
\item{new_url}{the new url to send requests.}
}
\description{
set the default location for all aws.s3
requests. By default, the base_url is "s3.amazonaws.com".
}
