% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bucketname}
\alias{get_bucketname}
\alias{get_bucketname.character}
\alias{get_bucketname.s3_bucket}
\alias{get_bucketname.s3_object}
\title{Utility Functions}
\usage{
get_bucketname(x, ...)

\method{get_bucketname}{character}(x, ...)

\method{get_bucketname}{s3_bucket}(x, ...)

\method{get_bucketname}{s3_object}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{\dots}{Ignored.}
}
\value{
\code{get_bucketname} returns a character string.
}
\description{
Some utility functions for working with S3 objects and buckets
}
