% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{write_object}
\alias{write_object}
\title{Write an object locally}
\usage{
write_object(bucket, object, con, headers = list(), ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{object}{Character string of the name of the object you want to get.}

\item{con}{An R connection (e.g., a character string specifying a file path).}

\item{headers}{List of request headers for the REST call.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A raw object, invisibly.
}
\description{
Download an S3 object and save it locally
}
\details{
This function is identical to \code{\link{getobject}}, except
in addition to returning the object as a raw vector within R, this function
writes the object to a connection, as specified by the \code{con} argument.
}
\examples{
\dontrun{}
}
\keyword{object}

