% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.s3-package.R
\docType{package}
\name{aws.s3-package}
\alias{aws.s3-package}
\alias{aws.s3}
\title{aws.s3-package}
\description{
AWS S3 Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) Simple Storage Service (S3) REST API.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
