% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_exists.R
\name{bucket_exists}
\alias{bucket_exists}
\title{Bucket exists?}
\usage{
bucket_exists(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if bucket exists and is accessible, else \code{FALSE}.
}
\description{
Check whether a bucket exists and is accessible with the current authentication keys.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketHEAD.html}{API Documentation}
}
\seealso{
\code{\link{bucketlist}}, \code{\link{get_bucket}}, \code{\link{object_exists}}
}
