% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERVICE.R
\name{bucketlist}
\alias{bucketlist}
\title{bucketlist}
\usage{
bucketlist(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
a list of buckets.  if passed with default settings \dQuote{parse_response = TRUE}
the response will list the name and creationdate of all buckets owned by the request
sender.
}
\description{
List buckets for authenticated user
}
\details{
\code{bucketlist} performs a GET operation on the base s3 endpoint 
and returns a list of all buckets owned by the authenticated 
sender of the request. If authentication is successful, this function provides a list of 
buckets available to the authenticated user. In this way, it can serve as 
a \dQuote{hello world!} function, to confirm that one's authentication 
credentials are working correctly.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTServiceGET.html}{API Documentation}
}
\keyword{service}

