% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketlist.R
\name{bucketlist}
\alias{bucketlist}
\title{List Buckets}
\usage{
bucketlist(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A data frame of buckets.
}
\description{
List buckets as a data frame
}
\details{
\code{bucketlist} performs a GET operation on the base s3 endpoint and returns a list of all buckets owned by the authenticated sender of the request. If authentication is successful, this function provides a list of buckets available to the authenticated user. In this way, it can serve as a \dQuote{hello world!} function, to confirm that one's authentication credentials are working correctly.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTServiceGET.html}{API Documentation}
}
\seealso{
\code{\link{get_bucket}}, \code{\link{get_object}}
}
\keyword{service}

