% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{copyobject}
\alias{copyobject}
\title{Copy an object into another S3 bucket}
\usage{
copyobject(from_object, to_object = from_object, from_bucket, to_bucket,
  headers = list(), ...)
}
\arguments{
\item{from_object}{A character string containing the name the object you want to copy.}

\item{from_bucket}{A character string containing the name of the bucket you want to copy from.}

\item{to_bucket}{A character string containing the name of the bucket you want to copy into.}

\item{headers}{List of request headers for the REST call.}

\item{...}{additional arguments passed to \code{\link{s3HTTP}}}

\item{from_object}{A character string containing the name the object should have in the new bucket.}
}
\value{
Something...
}
\description{
Copy an object into another S3 bucket
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectCOPY.html}{API Documentation}
}

