% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{delete_cors}
\alias{delete_cors}
\title{Delete the cross origin resource sharing configuration information for a bucket.}
\usage{
delete_cors(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise.
}
\description{
Delete the cross origin resource sharing configuration information for a bucket.
}

