% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{delete_cors}
\alias{delete_cors}
\title{Delete the cross origin resource sharing configuration information for a bucket.}
\usage{
delete_cors(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketDELETEcors.html}{API Documentation}
}

