% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{delete_policy}
\alias{delete_policy}
\title{Delete the bucket access policy for a bucket.}
\usage{
delete_policy(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise.
An \code{aws_error} object may be returned if the request failed.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketDELETEpolicy.html}{API Documentation}
}

