% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{deletebucket}
\alias{deletebucket}
\title{Deletes an S3 bucket.}
\usage{
deletebucket(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket you want to delete.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}}
}
\value{
TRUE if successful, aws_error details if not.
}
\description{
Deletes an S3 bucket.
}

