% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{deletebucket}
\alias{deletebucket}
\title{Deletes an S3 bucket.}
\usage{
deletebucket(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, \code{FALSE} otherwise.
An \code{aws_error} object may be returned if the request failed.
}
\description{
Deletes an S3 bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketDELETE.html}{API Documentation}
}

