% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT.R
\name{deleteobject}
\alias{deleteobject}
\title{Delete object}
\usage{
deleteobject(object, bucket, ...)
}
\arguments{
\item{object}{Character string with the object key, or an object of class \dQuote{s3_object}. In most cases, if \code{object} is specified as the latter, \code{bucket} can be omitted because the bucket name will be extracted from \dQuote{Bucket} slot in \code{object}.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
TRUE if successful, aws_error details if not.
}
\description{
Deletes an object from an S3 bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectDELETE.html}{API Documentation}
}

