% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_cors}
\alias{get_cors}
\title{Get the cross origin resource sharing configuration information for a bucket.}
\usage{
get_cors(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list with cors configuration and rules.
}
\description{
Get the cross origin resource sharing configuration information for a bucket.
}

