% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_lifecycle}
\alias{get_lifecycle}
\title{Get the lifecycle configuration information for a bucket.}
\usage{
get_lifecycle(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list with lifecycle configuration, if it has been configured.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGETlifecycle.html}{API Documentation}
}

