% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_replication}
\alias{get_replication}
\title{Get the replication configuration for a bucket.}
\usage{
get_replication(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the replication configuration, if one has been set.
}
\description{
Get the replication configuration for a bucket.
}

