% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_requestpayment}
\alias{get_requestpayment}
\title{Get the requestPayment subresource for a bucket.}
\usage{
get_requestpayment(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the requestPayment information, if set.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTrequestPaymentGET.html}{API Documentation}
}

