% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_requestpayment}
\alias{get_requestpayment}
\title{Get the requestPayment subresource for a bucket.}
\usage{
get_requestpayment(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the requestPayment information, if set.
}
\description{
Get the requestPayment subresource for a bucket.
}

