% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT.R
\name{get_torrent}
\alias{get_torrent}
\title{Retrieves a Bencoded dictionary (BitTorrent) for an object from an S3 bucket.}
\usage{
get_torrent(bucket, object, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{object}{Character string of the name of the object you want to get.}

\item{...}{additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
Something.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGETtorrent.html}{API Documentation}
}

