% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BUCKET.r
\name{get_uploads}
\alias{get_uploads}
\title{Get a list of multipart uploads for a bucket.}
\usage{
get_uploads(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the multipart upload information.
}
\description{
Get a list of multipart uploads for a bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadListMPUpload.html}{API Documentation}
}

