% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_object.R
\name{get_uploads}
\alias{get_uploads}
\title{Multipart uploads}
\usage{
get_uploads(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list containing the multipart upload information.
}
\description{
Get a list of multipart uploads for a bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadListMPUpload.html}{API Documentation}
}
