% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUCKET.R
\name{get_versions}
\alias{get_versions}
\title{Get versions of bucket objects.}
\usage{
get_versions(bucket, ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A list.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGETVersion.html}{API Documentation}
}

