% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{getobject}
\alias{getobject}
\title{Retrieves an object from an S3 bucket}
\usage{
getobject(bucket, object, headers = list(), ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket.}

\item{object}{Character string of the name of the object you want to get.}

\item{headers}{List of request headers for the REST call.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A raw object.
}
\description{
Retrieves an object from an S3 bucket
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation}
}

