% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{getobject}
\alias{getobject}
\title{retrieves an object from a S3 bucket}
\usage{
getobject(bucket, object, ...)
}
\arguments{
\item{bucket}{Character string of the name of the bucket you want to get.}

\item{object}{Character string of the name of the object you want to get.}

\item{...}{additional arguments passed to \code{\link{s3HTTP}}}
}
\value{
raw object
}
\description{
retrieves an object from a S3 bucket
}

