% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getobject.R
\name{get_object}
\alias{get_object}
\alias{head_object}
\alias{save_object}
\title{Get object}
\usage{
get_object(object, bucket, headers = list(), ...)

save_object(object, bucket, file, headers = list(), ...)

head_object(object, bucket, ...)
}
\arguments{
\item{object}{Character string with the object key, or an object of class \dQuote{s3_object}. In most cases, if \code{object} is specified as the latter, \code{bucket} can be omitted because the bucket name will be extracted from \dQuote{Bucket} slot in \code{object}.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{headers}{List of request headers for the REST call.}

\item{file}{An R connection, or file name specifying the local file to save the object into.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
If \code{file = NULL}, a raw object. Otherwise, a character string containing the file name that the object is saved to.
}
\description{
Retrieve an object from an S3 bucket
}
\details{
\code{get_object} retrieves an object into memory. \code{save_object} saves an object to a local file. \code{head_object} checks whether an object exists by executing an HTTP HEAD request; this can be useful for checking object headers such as \dQuote{content-length} or \dQuote{content-type}.
}
\examples{
\dontrun{
# get an object in memory
## create bucket
b <- put_bucket("myexamplebucket")
## save a dataset to the bucket
s3save(mtcars, bucket = b, object = "mtcars")
obj <- get_bucket(b)
## get the object in memory
x <- get_object(obj[[1]])
load(rawConnection(x))
"mtcars" \%in\% ls()

# save an object locally
y <- get_object(obj[[1]], file = object[[1]][["Key"]])
y \%in\% dir()

# return parts of an object
## use 'Range' header to specify bytes
get_object(object = obj[[1]], headers = list('Range' = 'bytes=1-120'))
}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation: GET Object}

\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectHEAD.html}{API Document: HEAD Object}
}
\seealso{
\code{\link{get_bucket}}
}

