% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_bucket.R
\name{put_bucket}
\alias{put_bucket}
\title{Create bucket}
\usage{
put_bucket(bucket, region = Sys.getenv("AWS_DEFAULT_REGION"), ...)
}
\arguments{
\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{region}{A character string containing the AWS region. If missing, defaults to value of environment variable \env{AWS_DEFAULT_REGION}.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
\code{TRUE} if successful, aws_error details if not.
}
\description{
Creates a new S3 bucket.
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketPUT.html}{API Documentation}
}

