% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_object.R
\name{put_object}
\alias{put_object}
\title{Put object}
\usage{
put_object(file, object, bucket, headers = list(), ...)
}
\arguments{
\item{file}{A character string containing the filename (or full path) of the file you want to upload to S3.}

\item{object}{A character string containing the name the object should 
have in S3 (i.e., its "object key"). If missing, the filename is used.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{headers}{List of request headers for the REST call.}

\item{\dots}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
If successful, \code{TRUE}, otherwise an aws_error object.
}
\description{
Puts an object into an S3 bucket
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPUT.html}{API Documentation}
}
\seealso{
\code{\link{delete_object}}
}

