% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3HTTP.R
\name{s3HTTP}
\alias{s3HTTP}
\title{S3 HTTP Requests}
\usage{
s3HTTP(verb = "GET", bucket = "", path = "", query = NULL,
  headers = list(), request_body = "", accelerate = FALSE,
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  session_token = Sys.getenv("AWS_SESSION_TOKEN"), parse_response = TRUE,
  check_region = TRUE, ...)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting to \dQuote{GET}.}

\item{bucket}{A character string with the name of the bucket, or an object of class \dQuote{s3_bucket}. If the latter and a region can be inferred from the bucket object attributes, then that region is used instead of \code{region}.}

\item{path}{A character string with the name of the object to put in the bucket (sometimes called the object or 'key name' in the AWS documentation.)}

\item{query}{Any query arguments, passed as a named list of key-value pairs.}

\item{headers}{A list of request headers for the REST call.}

\item{request_body}{A character string containing request body data.}

\item{accelerate}{A logical indicating whether to use AWS transfer acceleration, which can produce significant speed improvements for cross-country transfers. Acceleration only works with buckets that do not have dots in bucket name.}

\item{region}{A character string containing the AWS region. Ignored if region can be inferred from \code{bucket}. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key. If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}

\item{session_token}{Optionally, a character string containing an AWS temporary Session Token. If missing, defaults to value stored in environment variable \dQuote{AWS_SESSION_TOKEN}.}

\item{parse_response}{A logical indicating whether to return the response as is, or parse and return as a list. Default is \code{TRUE}.}

\item{check_region}{A logical indicating whether to check the value of \code{region} against the apparent bucket region. This is useful for avoiding (often confusing) out-of-region errors. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to an HTTP request function. such as \code{\link[httr]{GET}}.}
}
\value{
the S3 response, or the relevant error.
}
\description{
This is the workhorse function for executing API requests for S3.
}
\details{
This is mostly an internal function for executing API requests. In almost all cases, users do not need to access this directly.
}

