% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveload.r
\name{s3save}
\alias{s3load}
\alias{s3save}
\title{Save/load R object(s) to/from S3}
\usage{
s3save(..., bucket, object, opts)

s3load(bucket, object, opts, envir = parent.frame())
}
\arguments{
\item{...}{One or more R objects to be saved via \code{\link[base]{save}} and uploaded to S3.}

\item{bucket}{Character string with the name of the bucket.}

\item{object}{For \code{s3save}, a character string of the name of the object you want to save to. For \code{s3load}, a character string of the name of teh object you want to load from S3.}

\item{opts}{Additional arguments passed to \code{\link{s3HTTP}}.}

\item{envir}{An R environment to load objects into. Default is the \code{parent.frame()} from which the function is called.}
}
\value{
If successful, \code{NULL}, otherwise an error object.
}
\description{
Save/load R object(s) to/from S3
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation}
}

