% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveload.R
\name{s3save}
\alias{s3load}
\alias{s3save}
\title{Save/load}
\usage{
s3save(..., bucket, object, opts = NULL)

s3load(bucket, object, envir = parent.frame(), ...)
}
\arguments{
\item{...}{For \code{s3save}, one or more R objects to be saved via \code{\link[base]{save}} and uploaded to S3. For \code{s3load}, see \code{opts}.}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{object}{For \code{s3save}, a character string of the name of the object you want to save to. For \code{s3load}, a character string of the name of teh object you want to load from S3.}

\item{opts}{Additional arguments passed to \code{\link{s3HTTP}}.}

\item{envir}{An R environment to load objects into. Default is the \code{parent.frame()} from which the function is called.}
}
\value{
If successful, \code{NULL}, otherwise an error object.
}
\description{
Save/load R object(s) to/from S3
}
\examples{
\dontrun{
# create bucket
b <- putbucket("myexamplebucket")

# save a dataset to the bucket
s3save(mtcars, bucket = b, object = "mtcars")
getbucket(b)

# load the data from bucket
s3load(bucket = b, object = "mtcars")
}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation}
}

