% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{s3saveRDS}
\alias{s3readRDS}
\alias{s3saveRDS}
\title{Serialization Interface for Single Objects}
\usage{
s3saveRDS(object, bucket, object_name, opts = NULL)

s3readRDS(bucket, object_name, opts = NULL)
}
\arguments{
\item{object}{For \code{s3saveRDS}, one or more R objects to be saved via \code{\link[base]{saveRDS}} and uploaded to S3}

\item{bucket}{Character string with the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{object_name}{For \code{s3saveRDS}, a character string of the name of the object you want to save to. For \code{s3loadRDS}, a character string of the name of teh object you want to load from S3.}

\item{opts}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
for \code{s3readRDS}, an R object.
 For \code{s3saveRDS}, NULL invisibly.
}
\description{
Functions to read and write single R objects in Amazon S3.
}
\examples{
\dontrun{
# create bucket
b <- put_bucket("myexamplebucket")

# save a single object to s3
s3saveRDS(object = mtcars, bucket = b, object_name = "mtcars.rds")

# restore it under a different name
mtcars2 <- s3readRDS(bucket = b, object = "mtcars.rds")
identical(mtcars, mtcars2)
}
}
\references{
\href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation}
}

