% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OBJECT.r
\name{write_object}
\alias{write_object}
\title{Write an object locally}
\usage{
write_object(bucket, object, con, headers = list(), ...)
}
\arguments{
\item{bucket}{A character string containing the name of the bucket, or an object of class \dQuote{s3_bucket}.}

\item{object}{A character string containing the name of an object, or an object of class \dQuote{s3_object}.}

\item{con}{An R connection (e.g., a character string specifying a file path).}

\item{headers}{List of request headers for the REST call.}

\item{...}{Additional arguments passed to \code{\link{s3HTTP}}.}
}
\value{
A raw object, invisibly.
}
\description{
Download an S3 object and save it locally
}
\details{
This function is identical to \code{\link{getobject}}, except
in addition to returning the object as a raw vector within R, this function
writes the object to a connection, as specified by the \code{con} argument.
}
\examples{
\dontrun{}
}
\keyword{object}

